Vsphere - customizing interfaces and disks when cloning from template
After this patch, the clone_vm accepts interfaces and volumes options,
taking interface resp. volume objects to modify the ones defined in
the template. This allows to customize all the interfaces (not just
changing the network_label on one of them), as well as increasing the
size of the volumes and adding or removing them.