[openstack] reduce code duplication between specs
This adds an RSpec.shared_context which contains the code that the
`before :all` blocks of both specs need.

I've merged and simplified the setup logic of both specs. Most
prominently, the environment variable $DEBUG is scrapped. There are only
two execution modes now, controlled by the presence of Openstack
credentials in the environment variables.