[openstack] fix test run so that the volume_spec.rb is actually run
For reasons that are currently beyond me, the Rake::TestTask (i.e. `rake
test`) does not run `spec/fog/openstack/*_spec.rb` (even though they
match the pattern of `spec/**/*_spec.rb`).

However, Rake's test loader loads the spec files as part of running the
test task. This caused syntax errors in the volume_spec to show up,
leading me to believe that the test is actually run (when it is only
ever compiled).

The "fix" is to expand @dhague's workaround of a special task for his
identity_v3_spec.rb to include my spec, too.