Openstack Model base class
base class documents how the standard interface should look
like. This doesn't bring anything to the code, since abstract
classes are redundant in ruby, but it is nice to see what should
be the standard interface.

Base class also define the repeated methods initialize and save,
so we don't need to define them in subclasess.

project is storable to attribute, since not all results contain
project associated, it's good to to be able to store the project
used when processing the collection.