Strip port number from host header
Excon by default adds the default port to the host header for each request
e.g hostname.example.com:443

Some Cloud Providers use filter rules on HAProxy or similar to control API paths to different backends. In our case the ruleset required will increase if we need to filter for port number as well as host headers where this is not specified. 

rfc2616 specifies that the port number should be appended to the host header if it is not the default port. However, 443 IS the default port for https (as specified in rfc2818) Although I accept that this could be seen as ambiguous given that the rfc2818 refers to HTTP over TLS rather than HTTP as defined in rfc2616.

Basically what I'm trying to say (in a long winded rounadbout way..) is should we make ":omit_default_port" (as supported by Excon) the default for vCloud Director  connections? 
If so I think this PR covers it whilst allowing users to override manually if required.