[ovirt] support passing cloud-init data as custom script
Currently, the cloud-init integration with ovirt supports only
a sub-set of cloud-init commands. With `custom_script` support,
one can pass arbitrary cloud-init yaml to the managed host and
make sure the provisioned host will get it unchanged (which
is not the case when passing the user data currently,
as the rbovirt parses into ovirt-specific keys and ignored the keys it
doesn't know)

Example:

```
cloudinit = <<EOF
yum_repos:
    zoo:
        baseurl: https://example.com/yumrepo
        name: example-repo
        enabled: true
EOF

vm.start_with_cloudinit(:blocking => true,
                        :user_data => cloudinit,
                        :use_custom_script => true)
```

This way, the cloudinit will get to the VM in the same format it was
passed during the call.

Requires 'https://github.com/abenari/rbovirt/pull/121'