Remove tests for deprecated binary `#[]`
Fog originally allowed use of `Provider[:service]` to generate a new
instance of a service. This was deprecated since it was not portable.
Using `AWS[:compute]` makes it harder to switch providers since all the
classes needed renaming. Ideally `Compute[@provider]` would be used and
can be configured within an application.

The `fog` script ("binary") had a number of backwards-compatibility
specs have been testing that the old methods worked. These generated
deprecation warnings as they ran.

Whilst the behaviour is still intact, this removes the tests that are
creating deprecation noise since no one is likely to be using them.