Tweak CloudSigma testing schema
The test schema for volumes had `storage_type` as a `String` which was
not being returned by the mock implementation.

According to the API docs I checked, it is an allowed value (but the
returned fields are controlled by the request).

So this makes the type `Fog::Nullable::String` so the schema passes if
the key/value are absent.